java: updated pin and password getting

+++ java/maika/ATMWithDB/src/finalproject/DeleteAccount.java
@@ -93,6 +93,21 @@ public class DeleteAccount {
                     return;
                 }
 
+                JPasswordField passwordField = new JPasswordField();
+                int result = JOptionPane.showConfirmDialog(null, passwordField, "Enter admin password:", JOptionPane.OK_CANCEL_OPTION);
+
+                if (!(result == JOptionPane.OK_OPTION)) {
+                    JOptionPane.showMessageDialog(null, "Cancelled");
+                    return;
+                }
+
+                String adminPass = String.valueOf(passwordField.getPassword());
+
+                if (!Database.admin.get("admin").equals(adminPass)) {
+                    JOptionPane.showMessageDialog(null, "Wrong password");
+                    return;
+                }
+
                 try {
                     Database.deleteUser(user.id);
                     JOptionPane.showMessageDialog(null, "Account deleted.");
